add channel mixer logic
add destroy on stop

+++ Assets/Scripts/Action Framework/Frame Player/Runtime Data/OnPlayUpdate.cs
@@ -4,5 +4,5 @@ public struct OnPlayUpdate : IComponentData
 {
     public bool loop;
     public int currentFrame;
-    public float normlizedTime;
+    public double normlizedTime;
 }
